refactor: change directory & rename webhook event schemas file
