+++ integrations/sendgrid/src/webhook-events/handlers/deferred.ts
@@ -1,6 +1,6 @@
 import * as bp from '../../../.botpress'
 import { unixTimestampToUtcDatetime } from '../../misc/utils'
-import { DeferredEmailWebhook } from '../sendgrid-webhook-schemas'
+import { DeferredEmailWebhook } from '../schemas/emails'
 
 export const handleDeferredEvent = async ({ client }: bp.HandlerProps, event: DeferredEmailWebhook) => {
   return await client.createEvent({
